+++ dap-etl/extraction/src/main/scala/org/broadinstitute/monster/dap/EnvironmentExtractionPipeline.scala
@@ -1,6 +1,13 @@
 package org.broadinstitute.monster.dap
 
-import org.broadinstitute.monster.dap.common.{Args, ExtractionPipelineBuilder}
+import org.broadinstitute.monster.dap.common.{
+  Args,
+  ExtractionPipelineBuilder,
+  FilterDirective,
+  FilterOps,
+  HttpWrapper,
+  OkWrapper
+}
 import org.broadinstitute.monster.common.{PipelineBuilder, ScioApp}
 
 // Ignore IntelliJ, this is used to make the implicit parser compile.
