+++ dap-etl/extraction/src/main/scala/org/broadinstitute/monster/dap/common/FilterDirective.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.common
 
 sealed trait FilterOp {
   def op: String
