+++ dap-etl/extraction/src/main/scala/org/broadinstitute/monster/dap/common/HttpWrapper.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.common
 
 import okhttp3.Request
 import upack.Msg
