+++ dap-etl/extraction/src/test/scala/org/broadinstitute/monster/dap/ExtractionPipelineBuilderSpec.scala
@@ -2,7 +2,7 @@ package org.broadinstitute.monster.dap
 
 import java.time.OffsetDateTime
 import better.files.File
-import common.{Args, ExtractionPipelineBuilder}
+import common.{Args, ExtractionPipelineBuilder, FilterDirective, FilterOps}
 import org.broadinstitute.monster.common.PipelineBuilderSpec
 import upack._
